# pref: improve executeBatch by avoiding statement-by-statement execution