QUERY_FORCE_DESCRIBE_PORTAL shared the same value as QUERY_DISALLOW_BATCHING (128), so batching was effectively disabled.

For instance, simple insert batch for 100 rows improves from 15ms to 1ms (when testing against local postgresql). The same 100 rows take 15ms for non-batched insert.

Before
```
Benchmark                                                   (nrows)  Mode  Cnt      Score        Error   Units
InsertBatch.insertBatch                                         100  avgt   10     15,052 ±      2,501   ms/op
InsertBatch.insertBatch:·gc.alloc.rate                          100  avgt   10      3,408 ±      0,501  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                     100  avgt   10  53319,755 ±    333,930    B/op
InsertBatch.insertBatch:·gc.churn.PS_Eden_Space                 100  avgt   10      6,351 ±     30,365  MB/sec
InsertBatch.insertBatch:·gc.churn.PS_Eden_Space.norm            100  avgt   10  91929,951 ± 439509,058    B/op
InsertBatch.insertBatch:·gc.count                               100  avgt   10      1,000               counts
InsertBatch.insertBatch:·gc.time                                100  avgt   10      2,000                   ms
InsertBatch.insertExecute                                       100  avgt   10     14,656 ±      2,628   ms/op
InsertBatch.insertExecute:·gc.alloc.rate                        100  avgt   10      2,940 ±      0,425  MB/sec
InsertBatch.insertExecute:·gc.alloc.rate.norm                   100  avgt   10  44744,617 ±     72,532    B/op
InsertBatch.insertExecute:·gc.churn.PS_Eden_Space               100  avgt   10      6,331 ±     30,268  MB/sec
InsertBatch.insertExecute:·gc.churn.PS_Eden_Space.norm          100  avgt   10  90687,654 ± 433569,747    B/op
InsertBatch.insertExecute:·gc.churn.PS_Survivor_Space           100  avgt   10      0,071 ±      0,338  MB/sec
InsertBatch.insertExecute:·gc.churn.PS_Survivor_Space.norm      100  avgt   10   1012,205 ±   4839,266    B/op
InsertBatch.insertExecute:·gc.count                             100  avgt   10      1,000               counts
InsertBatch.insertExecute:·gc.time                              100  avgt   10      2,000                   ms
```

After:

```
Benchmark                                                 (nrows)  Mode  Cnt      Score        Error   Units
InsertBatch.insertBatch                                       100  avgt   10      1,006 ±      0,062   ms/op
InsertBatch.insertBatch:·gc.alloc.rate                        100  avgt   10     50,430 ±      2,981  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                   100  avgt   10  53217,819 ±      5,771    B/op
InsertBatch.insertBatch:·gc.churn.PS_Eden_Space               100  avgt   10     55,125 ±     36,402  MB/sec
InsertBatch.insertBatch:·gc.churn.PS_Eden_Space.norm          100  avgt   10  58251,483 ±  38141,313    B/op
InsertBatch.insertBatch:·gc.churn.PS_Survivor_Space           100  avgt   10      0,028 ±      0,081  MB/sec
InsertBatch.insertBatch:·gc.churn.PS_Survivor_Space.norm      100  avgt   10     28,787 ±     82,714    B/op
InsertBatch.insertBatch:·gc.count                             100  avgt   10     10,000               counts
InsertBatch.insertBatch:·gc.time                              100  avgt   10      5,000                   ms
InsertBatch.insertExecute                                     100  avgt   10     14,403 ±      1,150   ms/op
InsertBatch.insertExecute:·gc.alloc.rate                      100  avgt   10      2,810 ±      0,212  MB/sec
InsertBatch.insertExecute:·gc.alloc.rate.norm                 100  avgt   10  42344,836 ±     73,813    B/op
InsertBatch.insertExecute:·gc.churn.PS_Eden_Space             100  avgt   10      4,961 ±     23,719  MB/sec
InsertBatch.insertExecute:·gc.churn.PS_Eden_Space.norm        100  avgt   10  72817,778 ± 348135,430    B/op
InsertBatch.insertExecute:·gc.count                           100  avgt   10      1,000               counts
InsertBatch.insertExecute:·gc.time                            100  avgt   10      1,000                   ms
```