@ringerc 2015-10-07
Thanks very much for spotting and fixing this (and for the benchmark!) @vlsi, and to @davecramer for the evaluation and merge.